尽量不用包装类

+++ src/main/java/com/ghostchu/tracker/sapling/entity/Permissions.java
@@ -37,5 +37,5 @@ public class Permissions implements Serializable {
     /**
      * 所属权限组
      */
-    private Integer targetGroup;
+    private long targetGroup;
 }
